[libiodata] Fix regexp used for parsing hex values
The function iodata::output::prepare(...) replaces certain
characters with hex values, like \x22 for a quotation mark.
The function is called when data is saved to a file,
see iodata::output::output_record(...).

However, the regexp used for reading saved data from a file
looked for hex values encoded as \x{22}, which is incorrect,
causing hex values of form \x22 be interpreted as stray
backslashes instead, which aborts the parsing process.